requirements.txt
# Fine-tuning (MLX)
mlx>=0.22.0
mlx-lm>=0.21.0

# Data extraction
pymupdf>=1.25.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdownify>=0.14.0

# Data processing
pandas>=2.2.0
nltk>=3.9.0
tiktoken>=0.8.0

# RAG / Embeddings
chromadb>=0.6.0
sentence-transformers>=3.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Evaluation
rouge-score>=0.1.2
numpy>=1.26.0

# Notebooks
jupyterlab>=4.0.0
matplotlib>=3.9.0

# Gemini (sender research for email replies)
google-genai>=1.0.0

# Gmail API (send emails from UI)
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0

# Google Analytics Data API (GA4)
google-analytics-data>=0.20.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
rich>=13.0.0
httpx>=0.28.0
